Avoids nil images when Shrine 3 doesn't accept params for file_url
Shrine 3 returns `nil` when a derivative doesn't exist. It also returns `nil`
when we pass parameters. This is a fallback way of avoiding any `nil` returns.
However, notice that we can't pass `params` because Shrine just ignores it and
returns `nil` if it's present (it wasn't like that in Shrine 2).