Fixing Devise in some edge cases
Overriding Devise's session controller because, for some very misterious reason,
it doesn't work in production. It works in some machines, not others,
despite using the same containers and environment variables. It is a
very specific behavior caused by warden.authenticate!.

I spent more time than I'd like on this so I decide for this fix which
works as intended and I deem low risk. Devise's controller has been the
same for years so I don't expect this to break anytime soon.